검색기능 추가 , 관리자 페이지 완료

+++ mysite03/src/main/java/com/douzone/mysite/repository/BoardRepository.java
@@ -97,7 +97,7 @@ public class BoardRepository {
 		
 		return result;
 	}
-	public void updateHit(int no) {
+	public void updateHit(Long no) {
 	
 		System.out.println("updateHit");
 		sqlSession.update("board.updateHit", no);
